pyproject.toml
[project]
name = "titanic-machine-learning-from-disaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "hydra-core>=1.3.2",
    "isort>=8.0.1",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "pandas>=3.0.2",
    "ruff>=0.15.11",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "ruff>=0.15.11",
]
